+++ tools/hulkz-viewer/src/app/layout.rs
@@ -1,11 +1,10 @@
 use crate::model::StreamId;
-use eframe::{egui, Storage};
-use egui_dock::{DockState, NodeIndex};
-use tracing::warn;
+use eframe::egui;
+use egui_dock::{DockState, Node, NodeIndex, SurfaceIndex};
 
 use super::{
-    panels, ParameterPanelTab, PersistedUiState, TextPanelTab, ViewerApp, ViewerTab,
-    STORAGE_KEY_DOCK_STATE, STORAGE_KEY_UI_STATE,
+    panels,
+    state::{ParameterPanelTab, TextPanelTab, ViewerApp, ViewerTab},
 };
 
 pub(super) struct ViewerTabHost<'a> {
@@ -26,8 +25,6 @@ impl egui_dock::TabViewer for ViewerTabHost<'_> {
 
     fn ui(&mut self, ui: &mut egui::Ui, tab: &mut Self::Tab) {
         match tab {
-            ViewerTab::Discovery => panels::draw_discovery_panel(self.app, ui),
-            ViewerTab::Timeline => panels::draw_timeline_panel(self.app, ui),
             ViewerTab::Text(stream) => panels::draw_text_panel(self.app, ui, stream),
             ViewerTab::Parameters(panel) => panels::draw_parameters_panel(self.app, ui, panel),
         }
@@ -50,15 +47,6 @@ pub(super) fn ensure_stream_tab_exists(
     }
 }
 
-pub(super) fn ensure_timeline_tab_exists(dock_state: &mut DockState<ViewerTab>) {
-    let has_timeline = dock_state
-        .iter_all_tabs()
-        .any(|(_, tab)| matches!(tab, ViewerTab::Timeline));
-    if !has_timeline {
-        dock_state.push_to_focused_leaf(ViewerTab::Timeline);
-    }
-}
-
 pub(super) fn apply_overrides_to_primary_text_panel(
     dock_state: &mut DockState<ViewerTab>,
     source_expression: Option<&str>,
@@ -78,7 +66,7 @@ pub(super) fn highest_stream_id(dock_state: &DockState<ViewerTab>) -> StreamId {
         .iter_all_tabs()
         .filter_map(|(_, tab)| match tab {
             ViewerTab::Text(stream) => Some(stream.id),
-            _ => None,
+            ViewerTab::Parameters(_) => None,
         })
         .max()
         .unwrap_or(0)
@@ -89,7 +77,7 @@ pub(super) fn highest_parameter_panel_id(dock_state: &DockState<ViewerTab>) -> u
         .iter_all_tabs()
         .filter_map(|(_, tab)| match tab {
             ViewerTab::Parameters(panel) => Some(panel.id),
-            _ => None,
+            ViewerTab::Text(_) => None,
         })
         .max()
         .unwrap_or(0)
@@ -100,45 +88,95 @@ pub(super) fn initial_dock_state(
     default_parameter_panel: ParameterPanelTab,
 ) -> DockState<ViewerTab> {
     let mut dock_state = DockState::new(vec![ViewerTab::Text(default_stream)]);
-    let [stream_leaf, _] = dock_state.main_surface_mut().split_left(
-        NodeIndex::root(),
-        0.72,
-        vec![ViewerTab::Discovery],
-    );
-    let [stream_leaf, _] =
-        dock_state
-            .main_surface_mut()
-            .split_below(stream_leaf, 0.85, vec![ViewerTab::Timeline]);
     let _ = dock_state.main_surface_mut().split_right(
-        stream_leaf,
+        NodeIndex::root(),
         0.78,
         vec![ViewerTab::Parameters(default_parameter_panel)],
     );
     dock_state
 }
 
-pub(super) fn load_persisted_dock_state(
-    storage: Option<&dyn Storage>,
-) -> Option<DockState<ViewerTab>> {
-    let storage = storage?;
-    let raw = storage.get_string(STORAGE_KEY_DOCK_STATE)?;
-    match serde_json::from_str::<DockState<ViewerTab>>(&raw) {
-        Ok(state) => Some(state),
-        Err(error) => {
-            warn!(?error, "failed to deserialize persisted dock state");
-            None
+const MIN_SPLIT_FRACTION: f32 = 0.05;
+const MAX_SPLIT_FRACTION: f32 = 0.95;
+
+pub(super) fn sanitize_dock_splits(dock_state: &mut DockState<ViewerTab>) -> bool {
+    let mut changed = false;
+    let mut surface_index = 0usize;
+    loop {
+        let Some(surface) = dock_state.get_surface_mut(SurfaceIndex(surface_index)) else {
+            break;
+        };
+        for node in surface.iter_nodes_mut() {
+            if let Node::Vertical(split) | Node::Horizontal(split) = node {
+                let original = split.fraction;
+                let sanitized = if original.is_finite() {
+                    original.clamp(MIN_SPLIT_FRACTION, MAX_SPLIT_FRACTION)
+                } else {
+                    0.5
+                };
+                if (sanitized - original).abs() > f32::EPSILON {
+                    split.fraction = sanitized;
+                    changed = true;
+                }
+            }
         }
+        surface_index = surface_index.saturating_add(1);
     }
+    changed
 }
 
-pub(super) fn load_persisted_ui_state(storage: Option<&dyn Storage>) -> Option<PersistedUiState> {
-    let storage = storage?;
-    let raw = storage.get_string(STORAGE_KEY_UI_STATE)?;
-    match serde_json::from_str::<PersistedUiState>(&raw) {
-        Ok(state) => Some(state),
-        Err(error) => {
-            warn!(?error, "failed to deserialize persisted ui state");
-            None
+#[cfg(test)]
+mod tests {
+    use super::{
+        initial_dock_state, sanitize_dock_splits, ParameterPanelTab, TextPanelTab,
+        MAX_SPLIT_FRACTION, MIN_SPLIT_FRACTION,
+    };
+    use crate::app::state::ViewerTab;
+    use egui_dock::Node;
+
+    #[test]
+    fn sanitize_dock_splits_clamps_invalid_split_fractions() {
+        let mut dock_state = initial_dock_state(
+            TextPanelTab::new(0, "odometry".to_string()),
+            ParameterPanelTab::new(0),
+        );
+        for node in dock_state.main_surface_mut().iter_mut() {
+            if let Node::Vertical(split) | Node::Horizontal(split) = node {
+                split.fraction = 0.0;
+            }
         }
+
+        let changed = sanitize_dock_splits(&mut dock_state);
+        assert!(changed);
+        for node in dock_state.main_surface().iter() {
+            if let Node::Vertical(split) | Node::Horizontal(split) = node {
+                assert!((MIN_SPLIT_FRACTION..=MAX_SPLIT_FRACTION).contains(&split.fraction));
+            }
+        }
+    }
+
+    #[test]
+    fn sanitize_dock_splits_noop_for_valid_layout() {
+        let mut dock_state = initial_dock_state(
+            TextPanelTab::new(0, "odometry".to_string()),
+            ParameterPanelTab::new(0),
+        );
+        assert!(!sanitize_dock_splits(&mut dock_state));
+    }
+
+    #[test]
+    fn initial_layout_contains_workspace_tabs() {
+        let dock_state = initial_dock_state(
+            TextPanelTab::new(0, "odometry".to_string()),
+            ParameterPanelTab::new(0),
+        );
+        let tabs = dock_state
+            .iter_all_tabs()
+            .map(|(_, tab)| tab)
+            .collect::<Vec<_>>();
+        assert!(tabs.iter().any(|tab| matches!(tab, ViewerTab::Text(_))));
+        assert!(tabs
+            .iter()
+            .any(|tab| matches!(tab, ViewerTab::Parameters(_))));
     }
 }
