+++ tools/hulkz-viewer/src/app/persistence.rs
@@ -1,0 +1,59 @@
+use eframe::Storage;
+use egui_dock::DockState;
+use tracing::warn;
+
+use super::state::{PersistedUiState, ViewerApp, ViewerTab};
+
+pub(super) const STORAGE_KEY_DOCK_STATE: &str = "hulkz_viewer/workspace_dock_state_v4";
+pub(super) const STORAGE_KEY_UI_STATE: &str = "hulkz_viewer/ui_state_v3";
+
+pub(super) fn load_persisted_dock_state(
+    storage: Option<&dyn Storage>,
+) -> Option<DockState<ViewerTab>> {
+    let storage = storage?;
+    let raw = storage.get_string(STORAGE_KEY_DOCK_STATE)?;
+    match serde_json::from_str::<DockState<ViewerTab>>(&raw) {
+        Ok(state) => Some(state),
+        Err(error) => {
+            warn!(?error, "failed to deserialize persisted dock state");
+            None
+        }
+    }
+}
+
+pub(super) fn load_persisted_ui_state(storage: Option<&dyn Storage>) -> Option<PersistedUiState> {
+    let storage = storage?;
+    let raw = storage.get_string(STORAGE_KEY_UI_STATE)?;
+    match serde_json::from_str::<PersistedUiState>(&raw) {
+        Ok(state) => Some(state),
+        Err(error) => {
+            warn!(?error, "failed to deserialize persisted ui state");
+            None
+        }
+    }
+}
+
+pub(super) fn save_persisted_state(app: &ViewerApp, storage: &mut dyn Storage) {
+    match serde_json::to_string(&app.workspace.dock_state) {
+        Ok(json) => storage.set_string(STORAGE_KEY_DOCK_STATE, json),
+        Err(error) => {
+            warn!(?error, "failed to serialize dock state");
+            return;
+        }
+    }
+
+    let ui_state = PersistedUiState {
+        ingest_enabled: app.ui.ingest_enabled,
+        follow_live: app.ui.follow_live,
+        next_stream_id: app.workspace.next_stream_id,
+        next_parameter_panel_id: app.workspace.next_parameter_panel_id,
+        default_namespace: app.ui.default_namespace_input.clone(),
+        show_discovery: app.shell.show_discovery,
+        show_timeline: app.shell.show_timeline,
+    };
+
+    match serde_json::to_string(&ui_state) {
+        Ok(json) => storage.set_string(STORAGE_KEY_UI_STATE, json),
+        Err(error) => warn!(?error, "failed to serialize ui state"),
+    }
+}
